# Conflicts:
#	src/main/java/com/yoyomo/domain/mail/application/dto/MailRequest.java
#	src/main/java/com/yoyomo/domain/mail/domain/service/MailReserveService.java

+++ src/main/java/com/yoyomo/domain/application/application/usecase/ApplyUseCase.java
@@ -1,5 +1,8 @@
 package com.yoyomo.domain.application.application.usecase;
 
+import static com.yoyomo.domain.application.application.dto.request.ApplicationRequestDTO.Save;
+import static com.yoyomo.domain.application.application.dto.response.ApplicationResponseDTO.MyResponse;
+
 import com.yoyomo.domain.application.application.dto.request.ApplicationRequestDTO.Update;
 import com.yoyomo.domain.application.application.dto.response.ApplicationResponseDTO.Response;
 import com.yoyomo.domain.application.application.mapper.ApplicationMapper;
@@ -15,17 +18,12 @@ import com.yoyomo.domain.item.application.usecase.ItemManageUseCase;
 import com.yoyomo.domain.item.domain.entity.Item;
 import com.yoyomo.domain.recruitment.domain.entity.Recruitment;
 import com.yoyomo.domain.recruitment.domain.service.RecruitmentGetService;
-import com.yoyomo.domain.user.application.mapper.UserMapperImpl;
 import com.yoyomo.domain.user.domain.entity.User;
+import com.yoyomo.domain.user.domain.service.UserGetService;
 import jakarta.transaction.Transactional;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-import static com.yoyomo.domain.application.application.dto.request.ApplicationRequestDTO.Save;
-import static com.yoyomo.domain.application.application.dto.response.ApplicationResponseDTO.MyResponse;
-import static com.yoyomo.domain.user.application.dto.request.UserRequestDTO.Find;
 
 @Service
 @RequiredArgsConstructor
@@ -39,39 +37,47 @@ public class ApplyUseCase {
     private final AnswerGetService answerGetService;
     private final AnswerUpdateService answerUpdateService;
     private final ApplicationUpdateService applicationUpdateService;
-    private final UserMapperImpl userMapper;
     private final ItemManageUseCase itemManageUseCase; // todo 삭제
+    private final UserGetService userGetService;
+
 
     @Transactional
-    public void apply(Save dto, String recruitmentId) {
+    public void apply(Save dto, String recruitmentId, Long userId) {
         Recruitment recruitment = recruitmentGetService.find(recruitmentId);
         recruitment.checkAvailable();
 
+        User applicant = userGetService.find(userId);
+
         List<Item> items = itemManageUseCase.create(dto.answers());
-        Application application = dto.toApplication(recruitment);
+        Application application = dto.toApplication(recruitment, applicant);
 
         applicationSaveService.save(recruitment, application); //todo application.answerId 추가
         answerSaveService.save(items, application.getId());
     }
 
-    public List<Response> readAll(Find dto) {
-        User user = userMapper.from(dto);
-        return applicationGetService.findAll(user).stream()
+    public List<Response> readAll(Long userId) {
+        User applicant = userGetService.find(userId);
+        return applicationGetService.findAll(applicant).stream()
                 .map(applicationMapper::toResponse)
                 .toList();
     }
 
-    public MyResponse read(String applicationId) {
+    public MyResponse read(String applicationId, Long userId) {
+        User applicant = userGetService.find(userId);
         Application application = applicationGetService.find(applicationId);
+        application.checkAuthorization(applicant);
+
         Answer answer = answerGetService.findByApplicationId(application.getId());
 
         return applicationMapper.toMyResponse(application, answer);
     }
 
     @Transactional
-    public void update(String applicationId, Update dto) {
+    public void update(String applicationId, Update dto, Long userId) {
+        User applicant = userGetService.find(userId);
         Application application = applicationGetService.find(applicationId);
         Recruitment recruitment = recruitmentGetService.find(application.getRecruitmentId());
+        application.checkAuthorization(applicant);
         recruitment.checkAvailable();
 
         List<Item> items = itemManageUseCase.create(dto.answers());
@@ -79,8 +85,10 @@ public class ApplyUseCase {
     }
 
     @Transactional
-    public void delete(String applicationId) {
-        Application application = applicationGetService.find(applicationId); // todo 권한조회
+    public void delete(String applicationId, Long userId) {
+        User applicant = userGetService.find(userId);
+        Application application = applicationGetService.find(applicationId);
+        application.checkAuthorization(applicant);
         applicationUpdateService.delete(application);
     }
 }
